backend/requirements.txt
Flask==3.1.1
gunicorn==23.0.0
flask-cors==5.0.1
mysql-connector-python==9.3.0